+++ bin/sphinx.py
@@ -1,6 +1,11 @@
 import sphinxval.sphinx
 import argparse
-
+import logging
+import logging.config
+import os
+from sphinxval.utils import config as cfg
+import pathlib
+import json
 
 parser = argparse.ArgumentParser()
 parser.add_argument("--ModelList", type=str, default='', \
@@ -18,8 +23,28 @@ parser.add_argument("--RelativePathPlots", type=bool, default=True, \
         help=("Generate reports with relative paths for plots"))
 
 
+#Create logger
+logger = logging.getLogger(__name__)
+
+
+def setup_logging():
+    # Create the logs/ directory if it does not yet exist
+    if not os.path.exists(cfg.logpath):
+        os.mkdir(cfg.logpath)
+
+    config_file = pathlib.Path('sphinxval/log/log_config.json')
+    with open(config_file) as f_in:
+        config = json.load(f_in)
+    logging.config.dictConfig(config)
+
 
 args = parser.parse_args()
 
-sphinx_df = sphinxval.sphinx.validate(args.DataList, args.ModelList, top=args.TopDirectory, Resume=args.Resume)
-sphinxval.sphinx.report.report(None, args.RelativePathPlots, sphinx_dataframe=sphinx_df)
+setup_logging()
+
+try:
+    sphinx_df = sphinxval.sphinx.validate(args.DataList, args.ModelList, top=args.TopDirectory, Resume=args.Resume)
+    sphinxval.sphinx.report.report(None, args.RelativePathPlots, sphinx_dataframe=sphinx_df)
+
+except:
+    logger.exception('SPHINX failed with an exception.')
